requirements.txt
django==5.0.0
djangorestframework==3.15.0
django-environ==0.11.2
django-cors-headers==4.4.0
drf-spectacular==0.27.2
Pillow==10.4.0
djangorestframework-simplejwt==5.3.1
faker==29.0.0
django_cloudinary_storag==0.3.0
cloudinary==1.41.0

